requirements.txt
pandas
streamlit
plotly
Pillow
numpy